feat: interface 를 추가해봄
