+++ src/main/kotlin/nextstep/subway/path/domain/PathInterface.kt
@@ -1,0 +1,5 @@
+package nextstep.subway.path.domain
+
+interface PathInterface {
+    fun getShortestPath(startStationId: Long, arrivalStationId: Long, type: String): Path
+}
